feat: properly read specialisations from bootspec and use them in lzbt-tool
This fixes https://github.com/DeterminateSystems/bootspec/issues/147 by vendoring
the bootspec crate and changing the type of specialisations from `Generation` to
`BootJson`.
This means that we capture the full recursive JSON schema for specialisations,
including extensions, while the previous type threw this information away.

This is essential so that users can set separate sort keys for specialisations,
to avoid the boot loader falling back to specialisations when a new generation
fails to boot. Often specialisations are special cases that we don't want to boot
automatically, so they can now be sorted below all normal generations so that if
a generation fails to boot, the boot loader will fall back onto the previous generation.